Fix covid flow
